requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
Django==6.0.3
django-allauth==65.15.0
django-cors-headers==4.9.0
django-crispy-forms==2.6
django-jazzmin==3.0.4
djangorestframework==3.17.0
gunicorn==25.3.0
idna==3.11
packaging==26.0
pillow==12.1.1
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.2
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
